Revert "feat: Add valueOf() on enumerated classes"
This reverts commit 21f8c53fa38b9450b9380755eca0c5c291ffae32.

Enumerated types in KoLmafia previously used a custom `valueOf()`
implementation while still enjoying the benefits of string concatenation
. Unfortunately, this was unintentionally exploiting a bug in the Rhino
engine. Not only was this unreproducible in other JS runtimes, it could
be fixed in a future release of Rhino, which would break all JS scripts
that relied on this buggy behavior. To prevent this issue, KoLmafia
dropped the custom `valueOf()` implementation in r20620. We therefore
remove the type definitions for `valueOf()` as well.

See:

- kolmafia.us forum thread describing the issue in detail:
  https://kolmafia.us/threads/javascript-remove-custom-valueof-implementation-for-enumerated-types.25844/#post-161259
- KoLmafia r20620 post:
  https://kolmafia.us/threads/20620-remove-valueof-as-per-philmasterplus.25845/