feat!: Bump TypeScript version req to 4.1+
The type definition for `eudora(newEudora)` uses `Lowercase<T>`, which
was added in TypeScript 4.1. Because of this, we now require TS 4.1+.

TypeScript 4.1 was released in November 2020 (see
https://devblogs.microsoft.com/typescript/announcing-typescript-4-1/).
Hopefully, enough time has passed that asking users to adopt TS 4.1+ is
not a big problem--and if it is, time would solve it quickly.