feat!: Mark constructors of built-ins private
Mark the constructors of KoLmafia built-in classes (Item, Effect, ...)
as private. This prevents users from attempting to instantiate them
directly, which would otherwise cause KoLmafia's JavaScript runtime to
abort.

Also add test cases that verify whether it is a type error to:

1. Instantiate the class directly
2. Call the constructor as a function

---

Note: One alternative approach to forbidding instantiation is to replace
each class with a pair of interface and a namespace. The interface would
define the instance properties, and the namespace would define the
static properties.
However, this approach would prevent us from using `instanceof` checks
with the builtins, which is why I chose the current approach.