build: Add separate tsconfig for test script
The definitions for `require` and `module` clash with `@types/node`,
which is needed to run the test script (`run-tests.ts`).
To keep the setup working, separate `tsconfig.json` into two files:

- `tsconfig.json` is used to compile `src/` and `test-d/`.
- `tsconfig.test.json` is used to compiel `run-tests.ts`.

Also make ts-node use `tsconfig.test.json` for this.
(Hopefully, all this mess can be removed once #21 is merged.)